lambda/generator/requirements.txt
# langchain
langgraph
# langchain-openai
langchain-ollama
langchain-pinecone
python-dotenv
# langchainhub
langchain-community

# Push data to Pinecone
# pymupdf
# sentence-transformers
# langchain-huggingface

# lint
# black

# network related package
python-whois
dnspython

# Jupyter notebook support on VSCODE
# ipykernel

# Language detection
langdetect

# For training the triage model
# transformers
# accelerate
# bitsandbytes
# peft
# datasets
# trl
# pandas
# pyarrow